move comparation logic from IDEA class to Path
